refactor(programs): no-issue: extract program-export parser and make build hint explicit
- Move the self-contained xlsx-parsing unit out of formBuilder.js into a
  sibling parseProgramExport.js (owns WORKBOOK_FILE_EXTENSIONS, exports
  parseProgramExportWorkbook and loadProgramDefinitionFromUpload). The
  parser now does one thing and is independently unit-testable, and the
  AI routing module is ~160 lines lighter.
- Make the build correction hint a first-class `correctionHint` option of
  invokeProgramDefinitionBuild, appended in one controlled place, instead
  of concatenating onto the build input at the call site.
